[Day 9]

+++ demo/src/main/java/com/study/demo/controller/BoardController.java
@@ -2,14 +2,17 @@ package com.study.demo.controller;
 
 
 import com.study.demo.entity.Board;
+import com.study.demo.entity.User;
 import com.study.demo.service.JWTService;
+import com.study.demo.service.UserService;
+import io.jsonwebtoken.Claims;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpHeaders;
-import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
 import com.study.demo.service.BoardService;
+import javax.servlet.http.HttpServletRequest;
+import java.util.List;
 
 @Controller
 public class BoardController {
@@ -17,42 +20,71 @@ public class BoardController {
     @Autowired
     private BoardService boardService;
 
+    @Autowired
+    private UserService userService;
+    private JWTService jwtService = new JWTService();
 //    private JWTService jwtService = new JWTService();
 
     // 페이지 매핑 시 확장자(html) 제외하고 text 리턴
-    @GetMapping("/board")
-    public String boardWriteForm() {
+    @GetMapping("/board/write")
+    public String boardWritePage() {
 
         return "boardWrite";
     }
 
     @GetMapping("/board/list")
-    public String boardList(Model model) {
-        model.addAttribute("list", boardService.boardList());
+    public String boardListPage() {
         return "boardList";
     }
 
-    // 게시글 가져오는 POST 요청
-    @PostMapping("/board/add-board")
-    public String addBoardWrite(Board board) {
-
-        System.out.println(board.getId());
-        boardService.write(board);
-
-        return "";
+    @GetMapping("/board/get-list")
+    @ResponseBody
+    public List<Board> boardList(HttpServletRequest request) {
+        // 로그인 정보 확인
+        Claims token = jwtService.checkAuthorizationHeader(request);
+        if(token == null) return null;
+        // 토큰 내의 페이로드 값을 가져와야 한다.
+        return boardService.boardList();
     }
 
-    @GetMapping("/board/view")
+
+    // 게시글 가져오는 POST 요청
+    @PostMapping("/board/add-board")
+    @ResponseBody
+    public String addBoardWrite(@RequestBody Board board, HttpServletRequest request) {
+        // 로그인 정보 확인
+        System.out.println(board);
+        Claims token = jwtService.checkAuthorizationHeader(request);
+        if(token == null) return null;
+        // board 정보 DB에 추가
+        String userid = token.get("userId").toString();
+        System.out.println(userid);
+        User user = userService.getUser(userid);
+        System.out.println(user);
+        board.setUser(user);
+        try {
+            boardService.write(board);
+        }catch(Exception e){
+            System.out.println(e);
+            return null;
+        }
+        return "success";
+    }
+
+    @GetMapping("/board/view/{id}")
     // board/view?id=1 => 파라미터의 id로 들어감
-    public String boardView(Model model, Integer id) {
-        System.out.println(model);
+    public String boardView(Model model, @PathVariable("id") Integer id) {
+        System.out.println(id);
         model.addAttribute("board", boardService.boardView(id));
-        System.out.println(model);
         return "boardView";
     }
 
     @GetMapping("/board/delete")
-    public String deleteBoard(Integer id) {
+    public String deleteBoard(HttpServletRequest request, Integer id) {
+        // 로그인 정보 확인
+        Claims token = jwtService.checkAuthorizationHeader(request);
+        if(token == null) return null;
+
         boardService.delete(id);
 
         return "redirect:/board/list";
@@ -65,7 +97,9 @@ public class BoardController {
     }
 
     @PostMapping("/board/update/{id}")
-    public String updateBoard(@PathVariable("id") Integer id, Board board) {
+    public String updateBoard(HttpServletRequest request, @PathVariable("id") Integer id, Board board) {
+        Claims token = jwtService.checkAuthorizationHeader(request);
+        if(token == null) return null;
 
         Board boardTmp = boardService.boardView(id);
         boardTmp.setTitle(board.getTitle());
